* changed money variables from Int to Long * added non dividable balance amount support

+++ src/Balance.kt
@@ -1,4 +1,4 @@
-class Balance(val person: Person, var amount: Int) {
+class Balance(val person: Person, var amount: Long) {
     override fun equals(other: Any?) = other is Balance && person == other.person
     override fun hashCode() = person.hashCode()
     override fun toString(): String {
