+++ src/Transaction.kt
@@ -1,5 +1,5 @@
 import java.util.UUID
 
-class Transaction(val from: Person, val to: Person, val amount: Int) {
+class Transaction(val from: Person, val to: Person, val amount: Long) {
     val id: UUID = UUID.randomUUID()
 }
